+++ bot.py
@@ -25,10 +25,10 @@ class DiscordUptime(commands.Bot):
             intents=discord.Intents.default(),
             activity=discord.Activity(
                 type=getattr(discord.ActivityType, get_config("activity_type").lower()),
-                name="services",
+                name=get_config("activity_name"),
             ),
             help_command=DefaultHelpCommand()
-            if not get_config("disable_help") == "true"
+            if not get_config("disable_help")
             else None,
         )
         self.bot = bot